requirements.txt
streamlit
PyMuPDF
requests
python-docx
streamlit-analytics